+++ packages/best-parser/scripts/parse-real.ts
@@ -1,8 +1,8 @@
 import { readFile } from "node:fs/promises";
-import { getLogger } from "@emdzej/ediabasx-logger";
+import { getLogger } from "@emdzej/bimmerz-logger";
 import { parsePrg } from "../src/parser";
 
-const log = getLogger("best-parser.parse-real");
+const log = getLogger("EDIABASX.best-parser.parse-real");
 
 const filePath = process.argv[2];